+++ Tests/Unit/MiniAppScriptMessageHandlerTests.swift
@@ -1517,6 +1517,31 @@ class MiniAppScriptMessageHandlerTests: QuickSpec {
                     expect(callbackProtocol.errorMessage).toEventually(contain(MASDKError.failedToConformToProtocol.errorDescription ??  ""))
                 }
             }
+            context("when user controller receive isDarkMode action and id") {
+                let scriptMessageHandler = MiniAppScriptMessageHandler(
+                    delegate: callbackProtocol,
+                    hostAppMessageDelegate: mockMessageInterface,
+                    adsDisplayer: mockAdsDelegate,
+                    secureStorageDelegate: mockSecureStorageDelegate,
+                    miniAppManageDelegate: mockMiniAppManageInterface,
+                    miniAppId: mockMiniAppInfo.id,
+                    miniAppTitle: mockMiniAppTitle
+                )
+                it("will return true if the device is set to dark mode") {
+                    mockMessageInterface.mockIsDarkMode = true
+                    let mockMessage = MockWKScriptMessage(name: "isDarkMode", body: "{\"action\":\"isDarkMode\",\"param\":null,\"id\":\"9.408169489793706\"}" as AnyObject)
+                    scriptMessageHandler.userContentController(WKUserContentController(), didReceive: mockMessage)
+                    expect(callbackProtocol.messageId).toEventually(equal("9.408169489793706"))
+                    expect(callbackProtocol.response).toEventually(equal("true"))
+                }
+                it("will return false if the device is NOT set to dark mode") {
+                    mockMessageInterface.mockIsDarkMode = false
+                    let mockMessage = MockWKScriptMessage(name: "isDarkMode", body: "{\"action\":\"isDarkMode\",\"param\":null,\"id\":\"9.408169489793706\"}" as AnyObject)
+                    scriptMessageHandler.userContentController(WKUserContentController(), didReceive: mockMessage)
+                    expect(callbackProtocol.messageId).toEventually(equal("9.408169489793706"))
+                    expect(callbackProtocol.response).toEventually(equal("false"))
+                }
+            }
         }
     }
 }